Move kmsg collector to contrib
