+++ omnistat/monitor.py
@@ -99,7 +99,9 @@ class Monitor:
             "enable_rocprofiler", False
         )
 
-        self.runtimeConfig["collector_enable_kmsg"] = config["omnistat.collectors"].getboolean("enable_kmsg", False)
+        self.runtimeConfig["collector_enable_contrib_kmsg"] = config["omnistat.collectors"].getboolean(
+            "enable_contrib_kmsg", False
+        )
 
         allowed_ips = config["omnistat.collectors"].get("allowed_ips", "127.0.0.1")
         # convert comma-separated string into list
@@ -126,9 +128,11 @@ class Monitor:
         if config.has_option("omnistat.collectors.rocprofiler", "metrics"):
             self.runtimeConfig["rocprofiler_metrics"] = config["omnistat.collectors.rocprofiler"]["metrics"].split(",")
 
-        if config.has_section("omnistat.collectors.kmsg"):
-            self.runtimeConfig["kmsg_min_severity"] = config["omnistat.collectors.kmsg"].get("min_severity", "ERROR")
-            self.runtimeConfig["kmsg_include_existing"] = config["omnistat.collectors.kmsg"].getboolean(
+        if config.has_section("omnistat.collectors.contrib.kmsg"):
+            self.runtimeConfig["kmsg_min_severity"] = config["omnistat.collectors.contrib.kmsg"].get(
+                "min_severity", "ERROR"
+            )
+            self.runtimeConfig["kmsg_include_existing"] = config["omnistat.collectors.contrib.kmsg"].getboolean(
                 "include_existing_messages", False
             )
 
@@ -197,8 +201,8 @@ class Monitor:
                 rocprofiler(self.runtimeConfig["collector_rocm_path"], self.runtimeConfig["rocprofiler_metrics"])
             )
 
-        if self.runtimeConfig["collector_enable_kmsg"]:
-            from omnistat.collector_kmsg import KmsgCollector
+        if self.runtimeConfig["collector_enable_contrib_kmsg"]:
+            from omnistat.contrib.collector_kmsg import KmsgCollector
 
             min_severity = self.runtimeConfig["kmsg_min_severity"]
             include_existing = self.runtimeConfig["kmsg_include_existing"]
