+++ pyproject.toml
@@ -17,7 +17,7 @@ requires = ["setuptools>=61", "wheel", "setuptools-git-versioning>=2.0,<3"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["omnistat"]
+packages = ["omnistat", "omnistat.contrib"]
 
 [tool.setuptools.dynamic]
 dependencies = { file = ["requirements.txt"] }
